refactor: useStationSpeech refactor
